Refactor: Move shared FFI cdefs to separate header files
These are all the definitions that must be exported to LuaJIT, but are simultaneously used by the runtime itself (and type-checked).

No functional change on the C++ side of things, but it allows synchronizing the bindings more easily so that the LuaJIT version doesn't have to be updated manually any more.
