+++ Runtime/Bindings/runtime_ffi.hpp
@@ -2,13 +2,7 @@
 
 #include "lua.hpp"
 
-struct static_runtime_exports_table {
-	// Build configuration
-	const char* (*runtime_version)(void);
-
-	// REPL
-	void (*runtime_repl_start)(void);
-};
+#include "runtime_exports.h"
 
 namespace runtime_ffi {
 	void assignLuaState(lua_State* L);
